fix: #2 fixes auth response data types

+++ integration/src/main/java/pro/ra_tech/giga_ai_agent/integration/rest/model/AuthResponse.java
@@ -2,5 +2,5 @@ package pro.ra_tech.giga_ai_agent.integration.rest.model;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 
-public record AuthResponse(@JsonProperty("access_token") String accessToken, @JsonProperty("expires_at") int expiresAt) {
+public record AuthResponse(@JsonProperty("access_token") String accessToken, @JsonProperty("expires_at") long expiresAt) {
 }
